Update on toggle (!See description)
Unfortunately the toggle does not work correctly.
Toggling the hearing sensor to "on"  in "SoccerFieldTwos" does add the sensor to all the agents in the environment, can be confirmed by adding " Debug.Log("Sensor Name: " + sensors[i].GetName());" to /com.unity.ml-agents/Runtime/Agent.cs. However the HearingSensorComponent needs to be added via the unity editor to  the agents in order to by used by them. This component also needs to be removed when using the pretrained model by the ml-agents developers, the explanation can be viewed in the previous commit